% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parallelMatrixSqrt}
\alias{parallelMatrixSqrt}
\title{Compute in parallel the square root of matrix elements}
\usage{
parallelMatrixSqrt(x)
}
\arguments{
\item{orig}{A matrix}
}
\value{
A matrix in which each element is the square root of the corresponding element in the original matrix
}
\description{
Compute in parallel the square root of matrix elements
}
\examples{
m <- matrix(as.numeric(c(1:1000000)), nrow = 1000, ncol = 1000)
parallelMatrixSqrt(m)
}
